feat: replace Vercel OAuth with WorkOS AuthKit
Replace Vercel OAuth sign-in with WorkOS AuthKit to support Google
social login, GitHub social login, and email/password authentication
via WorkOS's hosted login UI.

- Add @workos-inc/node SDK and create WorkOS client module
- Create new /api/auth/signin/workos and /api/auth/workos/callback routes
- Delete Vercel OAuth routes, callback, and oauth.ts module
- Update session types (authProvider: "vercel" → "workos")
- Add "workos" to users.provider DB enum (text column, no migration needed)
- Simplify signout route (remove Vercel token revocation)
- Remove Vercel project sync UI from session starter
- Update sign-in button to generic "Sign in" with LogIn icon
- Update env vars, README, deploy docs, and skill docs
- Inline refreshVercelToken into vercel/token.ts (oauth.ts deleted)

Vercel sandbox infrastructure and GitHub App integration are unchanged.
Vercel token/project modules left in place (handle null gracefully).
