+++ apps/web/.env.example
@@ -6,10 +6,10 @@ JWE_SECRET=
 # Must be a 32-byte hex string (64 characters)
 ENCRYPTION_KEY=
 
-# Vercel OAuth (required for sign-in)
-# Callback URL: {YOUR_ORIGIN}/api/auth/vercel/callback
-NEXT_PUBLIC_VERCEL_APP_CLIENT_ID=
-VERCEL_APP_CLIENT_SECRET=
+# WorkOS AuthKit (required for sign-in)
+# Callback URL: {YOUR_ORIGIN}/api/auth/workos/callback
+WORKOS_API_KEY=
+WORKOS_CLIENT_ID=
 
 # GitHub App (required for repo access, pushes, and PRs)
 # Callback URL: {YOUR_ORIGIN}/api/github/app/callback
