+++ apps/web/app/[username]/[repo]/page.tsx
@@ -53,7 +53,7 @@ export default async function RepoPage({ params }: RepoPageProps) {
   const session = await getServerSession();
   if (!session?.user) {
     redirect(
-      `/api/auth/signin/vercel?next=${encodeURIComponent(`/${username}/${repo}`)}`,
+      `/api/auth/signin/workos?next=${encodeURIComponent(`/${username}/${repo}`)}`,
     );
   }
