+++ apps/web/app/api/auth/signin/github/route.ts
@@ -4,6 +4,6 @@ export async function GET(req: NextRequest): Promise<Response> {
   const next = req.nextUrl.searchParams.get("next") ?? "/";
   const params = new URLSearchParams({ next });
   return Response.redirect(
-    new URL(`/api/auth/signin/vercel?${params.toString()}`, req.url),
+    new URL(`/api/auth/signin/workos?${params.toString()}`, req.url),
   );
 }
