+++ apps/web/app/api/chat/route.test.ts
@@ -27,7 +27,7 @@ interface TestChatRecord {
 let sessionRecord: TestSessionRecord | null;
 let chatRecord: TestChatRecord | null;
 let currentAuthSession: {
-  authProvider?: "vercel" | "github";
+  authProvider?: "workos" | "github";
   user: {
     id: string;
     email?: string;
@@ -294,7 +294,7 @@ describe("/api/chat route", () => {
   test("blocks a sixth message for non-Vercel trial users on the managed deployment", async () => {
     const { POST } = await routeModulePromise;
     currentAuthSession = {
-      authProvider: "vercel",
+      authProvider: "workos",
       user: {
         id: "user-1",
         email: "person@example.com",
