+++ apps/web/app/api/github/app/callback/route.ts
@@ -149,7 +149,7 @@ export async function GET(req: Request): Promise<Response> {
 
   const session = await getServerSession();
   if (!session?.user?.id) {
-    const signinUrl = new URL("/api/auth/signin/vercel", req.url);
+    const signinUrl = new URL("/api/auth/signin/workos", req.url);
     signinUrl.searchParams.set(
       "next",
       `${requestUrl.pathname}${requestUrl.search}`,
