+++ apps/web/components/session-starter.tsx
@@ -9,12 +9,10 @@ import {
   Plus,
 } from "lucide-react";
 import Link from "next/link";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useGitHubConnectionStatus } from "@/hooks/use-github-connection-status";
 import { useSession } from "@/hooks/use-session";
 import { useUserPreferences } from "@/hooks/use-user-preferences";
-import { useVercelRepoProjects } from "@/hooks/use-vercel-repo-projects";
-import type { VercelProjectSelection } from "@/lib/vercel/types";
 import { cn } from "@/lib/utils";
 import { BranchSelectorCompact } from "./branch-selector-compact";
 import { RepoSelectorCompact } from "./repo-selector-compact";
@@ -23,7 +21,6 @@ import {
   SANDBOX_OPTIONS,
   type SandboxType,
 } from "./sandbox-selector-compact";
-import { SessionStarterVercelSyncSection } from "./session-starter-vercel-sync-section";
 import { Switch } from "./ui/switch";
 
 type SessionMode = "empty" | "repo";
@@ -38,7 +35,6 @@ interface SessionStarterProps {
     sandboxType: SandboxType;
     autoCommitPush: boolean;
     autoCreatePr: boolean;
-    vercelProject?: VercelProjectSelection | null;
   }) => void;
   isLoading?: boolean;
   lastRepo: { owner: string; repo: string } | null;
@@ -58,11 +54,8 @@ export function SessionStarter({
   const [selectedRepo, setSelectedRepo] = useState(() => lastRepo?.repo ?? "");
   const [selectedBranch, setSelectedBranch] = useState<string | null>(null);
   const [isNewBranch, setIsNewBranch] = useState(!!lastRepo);
-  const [vercelProjectChoice, setVercelProjectChoice] = useState<
-    string | null | undefined
-  >(undefined);
 
-  const { session, loading: sessionLoading, hasGitHub } = useSession();
+  const { hasGitHub } = useSession();
   const { reconnectRequired, isLoading: githubConnectionLoading } =
     useGitHubConnectionStatus({
       enabled: hasGitHub,
@@ -77,46 +70,11 @@ export function SessionStarter({
   const sandboxName =
     SANDBOX_OPTIONS.find((s) => s.id === sandboxType)?.name ?? sandboxType;
 
-  const shouldLoadVercelProjects =
-    mode === "repo" &&
-    !githubConnectionLoading &&
-    !reconnectRequired &&
-    !!selectedOwner &&
-    !!selectedRepo &&
-    session?.authProvider === "vercel";
-  const {
-    data: repoProjects,
-    loading: repoProjectsLoading,
-    error: repoProjectsError,
-  } = useVercelRepoProjects({
-    enabled: shouldLoadVercelProjects,
-    repoOwner: selectedOwner,
-    repoName: selectedRepo,
-  });
-
-  useEffect(() => {
-    if (!shouldLoadVercelProjects) {
-      setVercelProjectChoice(undefined);
-      return;
-    }
-    if (!repoProjects || repoProjectsLoading) return;
-    if (repoProjects.selectedProjectId) {
-      setVercelProjectChoice(repoProjects.selectedProjectId);
-      return;
-    }
-    if (repoProjects.projects.length === 0) {
-      setVercelProjectChoice(null);
-      return;
-    }
-    setVercelProjectChoice(undefined);
-  }, [repoProjects, repoProjectsLoading, shouldLoadVercelProjects]);
-
   const handleRepoSelect = (owner: string, repo: string) => {
     setSelectedOwner(owner);
     setSelectedRepo(repo);
     setSelectedBranch(null);
     setIsNewBranch(false);
-    setVercelProjectChoice(undefined);
   };
 
   const handleRepoClear = () => {
@@ -124,7 +82,6 @@ export function SessionStarter({
     setSelectedRepo("");
     setSelectedBranch(null);
     setIsNewBranch(false);
-    setVercelProjectChoice(undefined);
   };
 
   const handleBranchChange = (branch: string | null, newBranch: boolean) => {
@@ -139,54 +96,16 @@ export function SessionStarter({
 
   const isRepoSelectionComplete =
     mode !== "repo" || (selectedOwner && selectedRepo);
-  const isVercelLookupPending =
-    mode === "repo" &&
-    !!selectedOwner &&
-    !!selectedRepo &&
-    (sessionLoading || (shouldLoadVercelProjects && repoProjectsLoading));
-  const requiresVercelChoice =
-    shouldLoadVercelProjects &&
-    !repoProjectsLoading &&
-    !repoProjectsError &&
-    !!repoProjects &&
-    repoProjects.projects.length > 0 &&
-    repoProjects.selectedProjectId === null &&
-    vercelProjectChoice === undefined;
   const controlsDisabled = isLoading || preferencesLoading;
   const isSubmitDisabled =
     controlsDisabled ||
     (mode === "repo" && (githubConnectionLoading || reconnectRequired)) ||
-    !isRepoSelectionComplete ||
-    isVercelLookupPending ||
-    requiresVercelChoice;
+    !isRepoSelectionComplete;
   const effectiveAutoCommitPush = autoCommitPush ?? defaultAutoCommitPush;
   const effectiveAutoCreatePr = autoCreatePr ?? defaultAutoCreatePr;
-  const showVercelProjectSection =
-    mode === "repo" &&
-    !githubConnectionLoading &&
-    !reconnectRequired &&
-    !!selectedOwner &&
-    !!selectedRepo &&
-    (sessionLoading || session?.authProvider === "vercel");
 
   const handleSubmit = () => {
     if (isSubmitDisabled) return;
-
-    let vercelProject: VercelProjectSelection | null | undefined;
-    if (shouldLoadVercelProjects) {
-      if (repoProjectsError || !repoProjects) {
-        vercelProject = undefined;
-      } else if (vercelProjectChoice === null) {
-        vercelProject = null;
-      } else if (typeof vercelProjectChoice === "string") {
-        vercelProject =
-          repoProjects.projects.find(
-            (project) => project.projectId === vercelProjectChoice,
-          ) ?? null;
-      } else {
-        return;
-      }
-    }
 
     onSubmit({
       repoOwner: mode === "repo" ? selectedOwner || undefined : undefined,
@@ -200,7 +119,6 @@ export function SessionStarter({
       sandboxType,
       autoCommitPush: effectiveAutoCommitPush,
       autoCreatePr: effectiveAutoCommitPush ? effectiveAutoCreatePr : false,
-      vercelProject,
     });
   };
 
@@ -265,18 +183,6 @@ export function SessionStarter({
                   onChange={handleBranchChange}
                 />
               )}
-
-            {showVercelProjectSection && (
-              <SessionStarterVercelSyncSection
-                controlsDisabled={controlsDisabled}
-                isVercelLookupPending={isVercelLookupPending}
-                repoProjects={repoProjects}
-                repoProjectsError={repoProjectsError}
-                requiresVercelChoice={requiresVercelChoice}
-                vercelProjectChoice={vercelProjectChoice}
-                onVercelProjectChoiceChange={setVercelProjectChoice}
-              />
-            )}
           </div>
         )}
