+++ apps/web/hooks/use-sessions.ts
@@ -4,7 +4,6 @@ import { useCallback, useEffect } from "react";
 import { toast } from "sonner";
 import useSWR, { useSWRConfig } from "swr";
 import type { Chat, Session } from "@/lib/db/schema";
-import type { VercelProjectSelection } from "@/lib/vercel/types";
 import { fetcher } from "@/lib/swr";
 
 export type SessionWithUnread = Pick<
@@ -37,7 +36,6 @@ interface CreateSessionInput {
   sandboxType: "vercel";
   autoCommitPush: boolean;
   autoCreatePr: boolean;
-  vercelProject?: VercelProjectSelection | null;
 }
 
 interface SessionsResponse {
