+++ apps/web/lib/session/types.ts
@@ -1,6 +1,6 @@
 export interface Session {
   created: number;
-  authProvider: "vercel" | "github";
+  authProvider: "workos" | "github";
   user: {
     id: string;
     username: string;
@@ -12,7 +12,7 @@ export interface Session {
 
 export interface SessionUserInfo {
   user: Session["user"] | undefined;
-  authProvider?: "vercel" | "github";
+  authProvider?: "workos" | "github";
   hasGitHub?: boolean;
   hasGitHubAccount?: boolean;
   hasGitHubInstallations?: boolean;
