+++ apps/web/lib/workos/client.ts
@@ -1,0 +1,5 @@
+import "server-only";
+
+import { WorkOS } from "@workos-inc/node";
+
+export const workos = new WorkOS(process.env.WORKOS_API_KEY);
