+++ apps/web/package.json
@@ -38,6 +38,7 @@
     "@streamdown/code": "^1.0.3",
     "@vercel/analytics": "^1.4.1",
     "@vercel/oidc": "^2.0.0",
+    "@workos-inc/node": "^7.36.0",
     "ai": "catalog:",
     "arctic": "^3.7.0",
     "botid": "^1.5.11",
